POSSIBLY CONTROVERSIAL (you knew something like this was coming).
Add ability for GenomicRanges subclasses to list the names of slots
that should be treated as "columns" in the virtual GRanges table. This
is implemented by adding a method on the extraColumnSlotNames
generic. Then, the GenomicRanges methods take care of subsetting,
replacement, showing, some validity checking, etc, for those
slots. This greatly facilitates implementing subclasses with
additional fixed columns. Just overload extraColumnSlotNames, write
your constructor and slot accessors, and whatever else makes your
class special. Every attempt was made to implement this in a way that
does not negatively affect the performance of GRanges. If we had
shallow-R, this attention would not have been necessary.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicRanges@78284 bc3139a8-67e5-0310-9ffc-ced21a209358